+++ packages/midway-schedule/test/.setup.js
@@ -1,0 +1,1 @@
+process.env.PLUGIN_PATH = require('path').join(__dirname, '../../../');