+++ packages/midway-schedule/test/fixtures/app-load-schedule/src/lib/schedule/hello.ts
@@ -1,0 +1,13 @@
+'use strict';
+
+import { schedule } from '../../../../../../../midway';
+
+@schedule({
+  type: 'worker',
+  interval: 2333,
+})
+export default class HelloCron {
+  async exec(ctx) {
+    ctx.logger.info(process.pid, 'hello');
+  }
+}