+++ packages/midway-web/config/plugin.js
@@ -1,0 +1,6 @@
+module.exports = {
+  schedule: {
+    enable: true,
+    package: 'midway-schedule',
+  },
+};