+++ CHANGELOG.md
@@ -3,6 +3,35 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.1](https://github.com/midwayjs/midway/compare/v0.5.0...v0.5.1) (2018-11-15)
+
+
+### Bug Fixes
+
+* schedule build file not published ([4150ce2](https://github.com/midwayjs/midway/commit/4150ce2))
+
+
+
+
+
+# [0.5.0](https://github.com/midwayjs/midway/compare/v0.4.5...v0.5.0) (2018-11-15)
+
+
+### Bug Fixes
+
+* agent not work ([c7cf3a9](https://github.com/midwayjs/midway/commit/c7cf3a9))
+* agent startup become compatible between egg&midway ([05c98aa](https://github.com/midwayjs/midway/commit/05c98aa))
+* build not midway-bin ([5b9667f](https://github.com/midwayjs/midway/commit/5b9667f))
+* egg-schedule plugin retrieve dir ([8429332](https://github.com/midwayjs/midway/commit/8429332))
+* logger & build scripts ([ef1a948](https://github.com/midwayjs/midway/commit/ef1a948))
+* run task with wront ctx & fill tests ([30a0741](https://github.com/midwayjs/midway/commit/30a0741))
+
+
+### Features
+
+* init midway-schedule ([82cc9e1](https://github.com/midwayjs/midway/commit/82cc9e1))
+
+
 ## [0.4.7](https://github.com/midwayjs/midway/compare/v0.4.6...v0.4.7) (2018-11-15)